+++ src/Stella.Ergosfare.Core.Abstractions/Attributes/GroupAttribute.cs
@@ -1,4 +1,6 @@
-namespace Stella.Ergosfare.Contracts.Attributes;
+using System;
+
+namespace Stella.Ergosfare.Core.Abstractions.Attributes;
 
 
 /// <summary>
@@ -11,7 +13,7 @@ public class GroupAttribute(params string[] groupNames): Attribute
     /// The default group name used when no group is explicitly specified.
     /// </summary>
     public const string DefaultGroupName = "default";
-    
+
     /// <summary>
     /// Gets the group names assigned to this class.
     /// </summary>