+++ src/Stella.Ergosfare.Core/Extensions/TypeExtensions.cs
@@ -1,5 +1,5 @@
 ﻿using System.Diagnostics.CodeAnalysis;
-using Stella.Ergosfare.Contracts.Attributes;
+using Stella.Ergosfare.Core.Abstractions.Attributes;
 
 namespace Stella.Ergosfare.Core.Extensions;